Adds part I and II
This is where it begins. The book consists, as of now, of the notes
from the whole first and second courses, complete with examples and
some drawings. Though the first part has been proofread by others
and the second one is being slowly corrected, it's still far from
perfect: more examples and figures should also be added in time.