Corrects proof of Dini's theorem
Edited some bits in the proof of Dini's theorem, fixing the notation and
explaining some missing links.

Improved explanation of F_x and F_y in the generalized Dini theorem.

Also added brackets around partial derivatives in the formulae for the
second derivative of the implicit function for better reading.