Corrects proof of theorem of local invertibility
Added the name ("of local invertibility") of the theorem which states that
a function is a local diffeomorphism in a given point if its Jacobian
matrix in such point is non-singular.

Corrected some misplaced tildes over vectors and a scalar product
missing a bracket.

Corrected S^{n+1} (in Rn) to S^{n-1}, adding a footnote about spheres
and balls in Rn euclidean spaces.

Generally, rephrased the proof as it was deemed difficult.